Added settings

+++ FinalProject/Controllers/RecipeController.cs
@@ -6,20 +6,23 @@ using FinalProject.Models;
 using FinalProject.Repositories;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 
 namespace FinalProject.Controllers
 {
     public class RecipeController : Controller
     {
-        IRecipeRepository _recipeRepository;
-        IIngredientRepository _ingredientRepository;
-        IStepsRepository _stepsRepository;
+        private IRecipeRepository _recipeRepository;
+        private IIngredientRepository _ingredientRepository;
+        private IStepsRepository _stepsRepository;
+        private Settings _settings;
 
-        public RecipeController(IRecipeRepository recipeRepository, IIngredientRepository ingredientRepository, IStepsRepository stepsRepository)
+        public RecipeController(IOptions<Settings> settings, IRecipeRepository recipeRepository, IIngredientRepository ingredientRepository, IStepsRepository stepsRepository)
         {
             _recipeRepository = recipeRepository;
             _ingredientRepository = ingredientRepository;
             _stepsRepository = stepsRepository;
+            _settings = settings.Value;
         }
 
         [HttpGet]
@@ -41,36 +44,43 @@ namespace FinalProject.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create(RecipeModel model)
         {
-            for(int ii = 0; ii < model.Ingredients.Count; ii++)
+            if (_settings.AllowInsert)
             {
-                if (model.Ingredients[ii] == null)
+                for (int ii = 0; ii < model.Ingredients.Count; ii++)
                 {
-                    model.Ingredients.RemoveAt(ii);
-                    ii--;
+                    if (model.Ingredients[ii] == null)
+                    {
+                        model.Ingredients.RemoveAt(ii);
+                        ii--;
+                    }
                 }
-            }
-            for (int ii = 0; ii < model.Steps.Count; ii++)
-            {
-                if (model.Steps[ii] == null)
+                for (int ii = 0; ii < model.Steps.Count; ii++)
                 {
-                    model.Steps.RemoveAt(ii);
-                    ii--;
+                    if (model.Steps[ii] == null)
+                    {
+                        model.Steps.RemoveAt(ii);
+                        ii--;
+                    }
                 }
-            }
 
-            model.UserID = User.Claims.ElementAt(0).Value;
-            int newKey = _recipeRepository.Insert(model);
-            for (int ii = 0; ii < model.Ingredients.Count; ii++)
-            {
-                _ingredientRepository.Insert(newKey, ii + 1, model.Ingredients[ii]);
-            }
+                model.UserID = User.Claims.ElementAt(0).Value;
+                int newKey = _recipeRepository.Insert(model);
+                for (int ii = 0; ii < model.Ingredients.Count; ii++)
+                {
+                    _ingredientRepository.Insert(newKey, ii + 1, model.Ingredients[ii]);
+                }
 
-            for (int ii = 0; ii < model.Steps.Count; ii++)
-            {
-                _stepsRepository.Insert(newKey, ii + 1, model.Steps[ii]);
+                for (int ii = 0; ii < model.Steps.Count; ii++)
+                {
+                    _stepsRepository.Insert(newKey, ii + 1, model.Steps[ii]);
+                }
+
+                return RedirectToAction("DisplayRecipe", new { ID = newKey });
             }
-            
-            return RedirectToAction("DisplayRecipe", new { ID = newKey});
+            else
+            {
+                return RedirectToAction("Error", "Home");
+            }
         }
 
         [HttpGet]
@@ -91,13 +101,20 @@ namespace FinalProject.Controllers
         [Authorize(Roles = "Admin")]
         public IActionResult DeleteRecipe(int id)
         {
-            if (ModelState.IsValid)
+            if (_settings.AllowDelete)
             {
-                _ingredientRepository.Delete(id);
-                _stepsRepository.Delete(id);
-                _recipeRepository.Delete(id);
+                if (ModelState.IsValid)
+                {
+                    _ingredientRepository.Delete(id);
+                    _stepsRepository.Delete(id);
+                    _recipeRepository.Delete(id);
+                }
+                return RedirectToAction("Index", "Recipe");
             }
-            return RedirectToAction("Index", "Recipe");
+            else
+            {
+                return RedirectToAction("Error", "Home");
+            }
         }
 
         [HttpPost]
@@ -117,6 +134,5 @@ namespace FinalProject.Controllers
                 return View("SearchResults", null);
             }
         }
-
     }
 }
